+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 [Unreleased]
-* Fixed lines and bars drawing 
+* Fixed lines and bars drawing
+* Fixed missing "Read/Write" diagnostic recommendations
 
 ## [0.9.3-preview.2] - 2023-02-21
 * Added Texture mipmaps streaming analyzer