+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Added SRP Asset Settings analyzer.
 * Added Shader SRP Batcher analyzer.
 
 ## [0.9.3-preview.3] - 2023-02-28